REFACTOR:split files (useCartSync and useOrderData) into separate folders
