+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom';
 
 import { useAuthState } from './hooks/useAuthState';
 import { useGetProductsQuery } from './api/products/productApi';
-import { useCartSync } from './hooks/db/useCartSync';
+import { useCartSync } from './hooks/db/cart/useCartSync';
 
 import Divider from './components/atoms/Divider/Divider';
 import Header from './components/organisms/Header/Header';
